xcproj cleanup

+++ DFCache Tests/DFCache+Tests.m
@@ -17,7 +17,7 @@
 - (void)storeStringsWithCount:(NSUInteger)count strings:(NSDictionary *__autoreleasing *)s {
     NSMutableDictionary *strings = [NSMutableDictionary new];
     for (NSUInteger i = 0; i < count; i++) {
-        NSString *key = [NSString stringWithFormat:@"key_%i", i];
+        NSString *key = [NSString stringWithFormat:@"key_%lu", (unsigned long)i];
         NSString *string = [self _randomStringWithLength:(arc4random_uniform(30) + 1)];
         [self storeObject:string forKey:key cost:0 encode:DFCacheEncodeNSCoding];
         strings[key] = string;
